use double instead

+++ lib/model/settings/scan_duration_delegate.dart
@@ -2,11 +2,11 @@ import 'package:weforza/model/deferred_save_delegate.dart';
 
 /// This class represents the delegate that manages
 /// the scan duration option for the settings page.
-class ScanDurationDelegate extends DeferredSaveDelegate<int> {
+class ScanDurationDelegate extends DeferredSaveDelegate<double> {
   ScanDurationDelegate({required super.initialValue});
 
   @override
-  void saveValue(int value) {
+  void saveValue(double value) {
     // TODO: implement partial save of scan duration
   }
 }
